Restructure poeninja api and add ui
