+++ src/Sidekick.Business/Apis/Poe/Models/QueryResult.cs
@@ -1,3 +1,4 @@
+using Sidekick.Business.Apis.PoeNinja.Models;
 using Sidekick.Business.Parsers.Models;
 using System;
 using System.Collections.Generic;
@@ -15,5 +16,7 @@ namespace Sidekick.Business.Apis.Poe.Models
         public Item Item { get; set; }
 
         public Uri Uri { get; set; }
+
+        public PoeNinjaItem PoeNinjaItem { get; set; }
     }
 }
