+++ src/Sidekick.Business/Apis/PoeNinja/Models/PoeNinjaQueryResult.cs
@@ -1,10 +1,10 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Sidekick.Helpers.POENinjaAPI.Models
+namespace Sidekick.Business.Apis.PoeNinja.Models
 {
     public class PoeNinjaQueryResult<T> where T : PoeNinjaResult
     {
