+++ src/Sidekick.Business/Apis/PoeNinja/PoeNinjaClient.cs
@@ -1,6 +1,6 @@
 using Newtonsoft.Json;
 using Sidekick.Business.Http;
-using Sidekick.Helpers.POENinjaAPI.Models;
+using Sidekick.Business.Apis.PoeNinja.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -8,7 +8,7 @@ using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Sidekick.Helpers.POENinjaAPI
+namespace Sidekick.Business.Apis.PoeNinja
 {
     public class PoeNinjaClient
     {
