+++ src/Sidekick/Legacy.cs
@@ -1,4 +1,5 @@
 using Microsoft.Extensions.DependencyInjection;
+using Sidekick.Business.Apis.PoeNinja;
 using Sidekick.Business.Http;
 using Sidekick.Business.Languages.Client;
 using Sidekick.Business.Languages.UI;
@@ -43,6 +44,9 @@ namespace Sidekick
         public static ILeagueService LeagueService { get; private set; }
 
         [Obsolete]
+        public static PoeNinjaCache PoeNinjaCache { get; private set; }
+
+        [Obsolete]
         public static void Initialize()
         {
             InitializeService = Program.ServiceProvider.GetService<IInitializer>();
@@ -54,6 +58,7 @@ namespace Sidekick
             HttpClientProvider = Program.ServiceProvider.GetService<IHttpClientProvider>();
             LeagueService = Program.ServiceProvider.GetService<ILeagueService>();
             UILanguageProvider = Program.ServiceProvider.GetService<IUILanguageProvider>();
+            PoeNinjaCache = Program.ServiceProvider.GetService<PoeNinjaCache>();
         }
     }
 }
