+++ src/Sidekick/StartupExtensions.cs
@@ -1,4 +1,5 @@
 using Microsoft.Extensions.DependencyInjection;
+using Sidekick.Business.Apis.PoeNinja;
 using Sidekick.Core;
 using Sidekick.Services;
 
@@ -10,6 +11,13 @@ namespace Sidekick
         {
             services.AddInitializableService<ITrayService, TrayService>();
 
+            return services;
+        }
+
+        public static IServiceCollection AddPoeNinja(this IServiceCollection services)
+        {
+            services.AddSingleton<PoeNinjaClient>();
+            services.AddInitializableService<IPoeNinjaCache, PoeNinjaCache>();
             return services;
         }
     }
